Changed IncSearch color to orange
Search and IncSearch having the same colors, this makes it impossible to determine which higlighted search pattern is the active one.

I have changed the incSearch color from "Dracula green" to "Dracula orange".

This is an issue for example with substitute and the confirmation flag enabled: :%s/NONE/orange/cg