Add 'white' to palette
Fix #93 by removing bug from #92, which introduced code using the
'white' section of the palette. This did not exist.

I added it using the ANSI white definition of the spec, and then found
online that the closest 256-color is '255'. Open to change, but looks
very close (if not exact) to the screenshot in #92.