Add GitGutter/GitSigns highlighting
This adds specific highlight groups for GitGutter-related things.
It piggybacks off of the `Diff*` highlight groups.

The `gitsigns.nvim` maintainer [suggested](https://github.com/lewis6991/gitsigns.nvim/issues/94) that this be
implemented by checking if neovim is >= version 0.5, and if the package is
available (as opposed to a variable like `g:loaded_gitsigns`). This is because
`gitsigns` is essentially lazy-loaded so checking the variable when Dracula loads might be too early.

See these other related links in from [gitsigns](https://github.com/lewis6991/gitsigns.nvim):
https://github.com/lewis6991/gitsigns.nvim/issues/54
https://github.com/lewis6991/gitsigns.nvim/commit/317750d66a572588eef9a23fefce4aff1cbcad94