colors: support Saghen/blink.cmp
The plugin blink.cmp [1] has officially released version 1.0. With the
API now stable, Dracula can provide support for its highlight groups.
Define the blink item highlight groups (they're defined like the ones
for nvim-cmp). This allows users to use blink's highlight groups
directly instead of having to rely on the ones from nvim-cmp.

According to the docs from blink, the option to link the highlight
groups from blink to those from nvim-cmp is going to be removed in a
future release, so this additionally future proofs the Dracula theme
against this eventuality.

[1]: https://github.com/Saghen/blink.cmp